+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Upcoming
 
 - Add second dobra (STN) from São Tomé and Príncipe
+- Use euro symbol as html_entity for euro currency
+- Update Georgian Lari symbol
+- Add Ruby 3.1 to the CI matrix
 
 ## 6.18.0